requirements.txt
annotated-types
anyio
asgiref
boto3
botocore
certifi
cffi
charset-normalizer
cryptography
decorator
distro
Django
django-allauth
django-cors-headers
django-crispy-forms
django-filter
django-pwa
django-rest-auth
djangorestframework
et-xmlfile
fontawesomefree
fpdf
greenlet
h11
httpcore
httpx
idna
jiter
jmespath
numpy
openai
openpyxl
pandas
pillow
pycparser
pydantic
pydantic_core
PyJWT
python-dateutil
python-dotenv
pytz
regex
requests
s3transfer
six
sniffio
SQLAlchemy
sqlparse
tqdm
typing_extensions
tzdata
urllib3